Top Section 영양정보 서빙량에 따라 동적으로 계산되도록 수정
